refactor tests for mapping
