+++ src/main/java/org/opensearch/securityanalytics/action/CreateIndexMappingsRequest.java
@@ -8,15 +8,20 @@ import org.opensearch.action.ActionRequest;
 import org.opensearch.action.ActionRequestValidationException;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
+import org.opensearch.common.xcontent.ToXContentObject;
+import org.opensearch.common.xcontent.XContentBuilder;
 import org.opensearch.common.xcontent.XContentParser;
 import org.opensearch.common.xcontent.XContentParserUtils;
 
 import java.io.IOException;
+import java.util.Locale;
 
-public class CreateIndexMappingsRequest extends ActionRequest {
+import static org.opensearch.action.ValidateActions.addValidationError;
 
-    private static final String INDEX_NAME_FIELD = "index_name";
-    private static final String RULE_TOPIC_FIELD = "rule_topic";
+public class CreateIndexMappingsRequest extends ActionRequest implements ToXContentObject {
+
+    public static final String INDEX_NAME_FIELD = "index_name";
+    public static final String RULE_TOPIC_FIELD = "rule_topic";
 
     String indexName;
     String ruleTopic;
@@ -36,7 +41,14 @@ public class CreateIndexMappingsRequest extends ActionRequest {
 
     @Override
     public ActionRequestValidationException validate() {
-        return null;
+        ActionRequestValidationException validationException = null;
+        if (indexName == null || indexName.length() == 0) {
+            validationException = addValidationError(String.format(Locale.getDefault(), "%s is missing", INDEX_NAME_FIELD), validationException);
+        }
+        if (ruleTopic == null || ruleTopic.length() == 0) {
+            validationException = addValidationError(String.format(Locale.getDefault(), "%s is missing", RULE_TOPIC_FIELD), validationException);
+        }
+        return validationException;
     }
 
     @Override
@@ -92,5 +104,13 @@ public class CreateIndexMappingsRequest extends ActionRequest {
 
     public void setIndexName(String indexName) {
         this.indexName = indexName;
+    }
+
+    @Override
+    public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
+        return builder.startObject()
+                .field(INDEX_NAME_FIELD, indexName)
+                .field(RULE_TOPIC_FIELD, ruleTopic)
+                .endObject();
     }
 }
