+++ src/main/java/org/opensearch/securityanalytics/resthandler/RestUpdateIndexMappingsAction.java
@@ -6,6 +6,7 @@ package org.opensearch.securityanalytics.resthandler;
 
 import org.opensearch.client.node.NodeClient;
 import org.opensearch.common.xcontent.XContentParser;
+import org.opensearch.common.xcontent.XContentParserUtils;
 import org.opensearch.rest.BaseRestHandler;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.rest.action.RestToXContentListener;
@@ -32,13 +33,14 @@ public class RestUpdateIndexMappingsAction extends BaseRestHandler {
         UpdateIndexMappingsRequest req;
         if (!request.hasContentOrSourceParam()) {
             req = new UpdateIndexMappingsRequest(
-                    request.param("indexName"),
-                    request.param("field"),
-                    request.param("alias")
+                    request.param(UpdateIndexMappingsRequest.INDEX_NAME_FIELD),
+                    request.param(UpdateIndexMappingsRequest.FIELD),
+                    request.param(UpdateIndexMappingsRequest.ALIAS)
             );
         } else {
-            try (XContentParser parser = request.contentOrSourceParamParser()) {
-                req = UpdateIndexMappingsRequest.parse(parser);
+            try (XContentParser xcp = request.contentParser()) {
+                XContentParserUtils.ensureExpectedToken(XContentParser.Token.START_OBJECT, xcp.nextToken(), xcp);
+                req = UpdateIndexMappingsRequest.parse(xcp);
             }
         }
