+++ src/test/java/org/opensearch/securityanalytics/SecurityAnalyticsRestTestCase.java
@@ -9,25 +9,41 @@ import org.apache.http.HttpEntity;
 import org.apache.http.entity.ContentType;
 import org.apache.http.entity.StringEntity;
 import org.junit.Assert;
+import org.opensearch.action.admin.indices.mapping.get.GetMappingsResponse;
+import org.opensearch.action.search.SearchResponse;
 import org.opensearch.client.Request;
 import org.opensearch.client.RequestOptions;
 import org.opensearch.client.Response;
 import org.opensearch.client.RestClient;
 import org.opensearch.client.WarningsHandler;
+import org.opensearch.cluster.ClusterModule;
+import org.opensearch.cluster.metadata.MappingMetadata;
+import org.opensearch.common.collect.ImmutableOpenMap;
 import org.opensearch.common.settings.Settings;
+import org.opensearch.common.xcontent.DeprecationHandler;
+import org.opensearch.common.xcontent.NamedXContentRegistry;
 import org.opensearch.common.xcontent.ToXContent;
 import org.opensearch.common.xcontent.XContentBuilder;
 import org.opensearch.common.xcontent.XContentFactory;
+import org.opensearch.common.xcontent.XContentParser;
+import org.opensearch.common.xcontent.XContentParserUtils;
+import org.opensearch.common.xcontent.json.JsonXContent;
 import org.opensearch.commons.alerting.util.IndexUtilsKt;
+import org.opensearch.index.mapper.MapperService;
 import org.opensearch.rest.RestStatus;
+import org.opensearch.securityanalytics.action.CreateIndexMappingsRequest;
+import org.opensearch.securityanalytics.action.UpdateIndexMappingsRequest;
 import org.opensearch.securityanalytics.model.Detector;
 import org.opensearch.test.rest.OpenSearchRestTestCase;
 
 import java.io.IOException;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+
+import static org.opensearch.action.admin.indices.create.CreateIndexRequest.MAPPINGS;
 
 public class SecurityAnalyticsRestTestCase extends OpenSearchRestTestCase {
 
@@ -79,8 +95,64 @@ public class SecurityAnalyticsRestTestCase extends OpenSearchRestTestCase {
         return response;
     }
 
+    public static GetMappingsResponse executeGetMappingsRequest(String indexName) throws IOException {
+
+        Request getMappingsRequest = new Request("GET", indexName + "/_mapping");
+        Response response = client().performRequest(getMappingsRequest);
+
+        XContentParser parser = JsonXContent.jsonXContent.createParser(
+                new NamedXContentRegistry(ClusterModule.getNamedXWriteables()),
+                DeprecationHandler.THROW_UNSUPPORTED_OPERATION,
+                response.getEntity().getContent()
+        );
+        if (parser.currentToken() == null) {
+            parser.nextToken();
+        }
+
+        XContentParserUtils.ensureExpectedToken(parser.currentToken(), XContentParser.Token.START_OBJECT, parser);
+
+        Map<String, Object> parts = parser.map();
+
+        Map<String, MappingMetadata> mappings = new HashMap<>();
+        for (Map.Entry<String, Object> entry : parts.entrySet()) {
+            String _indexName = entry.getKey();
+            assert entry.getValue() instanceof Map : "expected a map as type mapping, but got: " + entry.getValue().getClass();
+
+            @SuppressWarnings("unchecked") final Map<String, Object> fieldMappings = (Map<String, Object>) ((Map<String, ?>) entry.getValue()).get(
+                    MAPPINGS.getPreferredName()
+            );
+
+            mappings.put(_indexName, new MappingMetadata(MapperService.SINGLE_MAPPING_NAME, fieldMappings));
+        }
+        ImmutableOpenMap<String, MappingMetadata> immutableMappingsMap =
+                new ImmutableOpenMap.Builder<String, MappingMetadata>().putAll(mappings).build();
+        return new GetMappingsResponse(immutableMappingsMap);
+    }
+
+    public static SearchResponse executeSearchRequest(String indexName, String queryJson) throws IOException {
+
+        Request request = new Request("GET", indexName + "/_search");
+        request.setJsonEntity(queryJson);
+        Response response = client().performRequest(request);
+
+        XContentParser parser = JsonXContent.jsonXContent.createParser(
+                new NamedXContentRegistry(ClusterModule.getNamedXWriteables()),
+                DeprecationHandler.THROW_UNSUPPORTED_OPERATION,
+                response.getEntity().getContent()
+        );
+        return SearchResponse.fromXContent(parser);
+    }
+
     protected HttpEntity toHttpEntity(Detector detector) throws IOException {
         return new StringEntity(toJsonString(detector), ContentType.APPLICATION_JSON);
+    }
+
+    protected HttpEntity toHttpEntity(CreateIndexMappingsRequest request) throws IOException {
+        return new StringEntity(toJsonString(request), ContentType.APPLICATION_JSON);
+    }
+
+    protected HttpEntity toHttpEntity(UpdateIndexMappingsRequest request) throws IOException {
+        return new StringEntity(toJsonString(request), ContentType.APPLICATION_JSON);
     }
 
     protected RestStatus restStatus(Response response) {
@@ -94,5 +166,15 @@ public class SecurityAnalyticsRestTestCase extends OpenSearchRestTestCase {
     private String toJsonString(Detector detector) throws IOException {
         XContentBuilder builder = XContentFactory.jsonBuilder();
         return IndexUtilsKt.string(shuffleXContent(detector.toXContent(builder, ToXContent.EMPTY_PARAMS)));
+    }
+
+    private String toJsonString(CreateIndexMappingsRequest request) throws IOException {
+        XContentBuilder builder = XContentFactory.jsonBuilder();
+        return IndexUtilsKt.string(shuffleXContent(request.toXContent(builder, ToXContent.EMPTY_PARAMS)));
+    }
+
+    private String toJsonString(UpdateIndexMappingsRequest request) throws IOException {
+        XContentBuilder builder = XContentFactory.jsonBuilder();
+        return IndexUtilsKt.string(shuffleXContent(request.toXContent(builder, ToXContent.EMPTY_PARAMS)));
     }
 }
