+++ src/test/java/org/opensearch/securityanalytics/action/GetIndexMappingsResponseTests.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.securityanalytics.mapper.action.mapping;
+package org.opensearch.securityanalytics.action;
 
 import org.opensearch.cluster.metadata.MappingMetadata;
 import org.opensearch.common.collect.ImmutableOpenMap;
